Remove duplication among acceptance controllers
