+++ app/controllers/schools/placement_requests/acceptance/preview_confirmation_email_controller.rb
@@ -2,8 +2,7 @@ module Schools
   module PlacementRequests
     module Acceptance
       class PreviewConfirmationEmailController < Schools::BaseController
-        before_action :set_placement_request
-        before_action :fetch_gitis_contact_for_placement_request
+        include Acceptance
         before_action :ensure_previous_step_complete
 
         def new; end
@@ -29,16 +28,6 @@ module Schools
         def candidate_booking_notification(booking)
           NotifyEmail::CandidateBookingConfirmation
             .from_booking(booking.candidate_email, booking.candidate_name, booking, candidates_cancel_url(booking.token))
-        end
-
-        def set_placement_request
-          @placement_request = @current_school
-            .bookings_placement_requests
-            .find(params[:placement_request_id])
-        end
-
-        def fetch_gitis_contact_for_placement_request
-          @placement_request.fetch_gitis_contact gitis_crm
         end
       end
     end
